requirements_full.txt
# Installation complète avec toutes les fonctionnalités
# Inclut le support PDF et les outils de visualisation

# Dépendances principales
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
owlready2>=0.48
rdflib>=6.3.2

# Dépendances pour fonctionnalités complètes
Pillow>=10.0.0           # Affichage d'images et traitement PDF
matplotlib>=3.7.0        # Graphiques statiques
plotly>=5.17.0          # Graphiques interactifs
openpyxl>=3.1.0         # Support fichiers Excel
pymupdf>=1.23.0         # Traitement PDF avancé

# Note: Pour le support PDF complet, vous pouvez aussi installer poppler:
# Sur Windows: choco install poppler ou télécharger depuis https://github.com/oschwartz10612/poppler-windows
# Sur Linux/macOS: apt-get install poppler-utils ou brew install poppler
